requirements.txt
Flask==3.0.0
openai==1.6.1
pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
google-cloud-logging ==3.9.0
